You are a world-class full-stack/src/user_auth/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
pydantic==2.4.2
pydantic[email]==2.4.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9